pyproject.toml
[project]
name = "homeassistant-grandstream-gwn"
version = "0.1.0"
description = "Integration for the Grandstream GWN Manager for Home Assistant"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    "homeassistant>=2025.4.4",
    "aiomqtt",
    "PyYAML",
]

[dependency-groups]
dev = [
    "mypy>=1.20.1",
    "ruff>=0.15.10",
    "types-PyYAML",
]

